Reorganize pages
