+++ .eslintrc.js
@@ -12,7 +12,14 @@ module.exports = {
     project: './tsconfig.json',
   },
   plugins: ['@typescript-eslint', 'prettier', 'react', 'jest'],
-  extends: ['react-app', 'prettier', 'prettier/react', 'prettier/@typescript-eslint', 'plugin:prettier/recommended'],
+  extends: [
+    'react-app',
+    'react-app/jest',
+    'prettier',
+    'prettier/react',
+    'prettier/@typescript-eslint',
+    'plugin:prettier/recommended',
+  ],
   env: {
     browser: true,
     es6: true,
