+++ src/navigation/logout.tsx
@@ -1,5 +1,5 @@
 import React, { FC } from 'react';
-import LogoutPage from '../components/LogoutPage';
+import LogoutPage from '../pages/LogoutPage';
 
 export const LogoutRoute: FC = () => {
   return <LogoutPage />;
